requirements.txt
requests
playwright
mysql-connector-python
python-dotenv